atualizações part.2
